Move `BRISKSignature512` to feature types header
