+++ common/src/feature_types.cpp
@@ -188,4 +188,13 @@ namespace pcl
     os << (i == 0 ? "(" : "") << p.histogram[i] << (i < (pcl::GFPFHSignature16::descriptorSize () - 1) ? ", " : ")");
     return (os);
   }
+
+  std::ostream&
+  operator << (std::ostream& os, const BRISKSignature512& p)
+  {
+    os << p.scale << " " << p.orientation << " ";
+    for (int i = 0; i < 64; ++i)
+    os << (i == 0 ? "(" : "") << p.descriptor[i] << (i < 63 ? ", " : ")");
+    return (os);
+  }
 } // namespace pcl
