+++ common/src/point_types.cpp
@@ -233,15 +233,6 @@ namespace pcl
   }
 
   std::ostream& 
-  operator << (std::ostream& os, const BRISKSignature512& p)
-  {
-    os << p.scale << " " << p.orientation << " ";
-    for (int i = 0; i < 64; ++i)
-    os << (i == 0 ? "(" : "") << p.descriptor[i] << (i < 63 ? ", " : ")");
-    return (os);
-  }
-
-  std::ostream& 
   operator << (std::ostream& os, const Narf36& p)
   {
     os << p.x<<","<<p.y<<","<<p.z<<" - "<<p.roll*360.0/M_PI<<"deg,"<<p.pitch*360.0/M_PI<<"deg,"<<p.yaw*360.0/M_PI<<"deg - ";
